Add `#[warn(clippy::cast_possible_truncation)]` to `console` crates

+++ console/types/string/src/random.rs
@@ -35,12 +35,12 @@ mod tests {
 
     type CurrentEnvironment = Console;
 
-    const ITERATIONS: u64 = 100;
+    const ITERATIONS: usize = 100;
 
     #[test]
     fn test_random() {
         // Initialize a set to store all seen random elements.
-        let mut set = HashSet::with_capacity(ITERATIONS as usize);
+        let mut set = HashSet::with_capacity(ITERATIONS);
 
         let mut rng = TestRng::default();
